+++ CHANGELOG.md
@@ -4,9 +4,13 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [1.2.0] - 2019-06-10
+### Added
+- Support TypedArray fields in RecordArray
+
 ## [1.1.1] - 2019-06-09
 ### Added
-- RecordArray#fromObject
+- RecordArray#fromObject ([b263...](https://github.com/zandaqo/structurae/commit/b263dc7503343a9114233e0a0ce027d6504d97cf))
 
 ## [1.1.0] - 2019-06-07
 ### Added